+++ password-protected.php
@@ -3,13 +3,12 @@
 Plugin Name: Password Protected
 Plugin URI: https://wordpress.org/plugins/password-protected/
 Description: A very simple way to quickly password protect your WordPress site with a single password. Please note: This plugin does not restrict access to uploaded files and images and does not work with some caching setups.
-Version: 2.5.1
+Version: 2.6.0
 Author: WPExperts
 Text Domain: password-protected
 Author URI: https://wpexperts.io/
 License: GPLv2
 */
-
 /*
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License, version 2, as
@@ -39,7 +38,7 @@ $Password_Protected = new Password_Protected();
 
 class Password_Protected {
 
-	var $version = '2.5.1';
+	var $version = '2.6.0';
 	var $admin   = null;
 	var $errors  = null;
 
@@ -69,6 +68,9 @@ class Password_Protected {
 		add_action( 'password_protected_login_messages', array( $this, 'login_messages' ) );
 		add_action( 'login_enqueue_scripts', array( $this, 'load_theme_stylesheet' ), 5 );
 
+		add_action('password_protected_above_password_field', array( $this, 'password_protected_above_password_field' ));
+		add_action('password_protected_below_password_field', array( $this, 'password_protected_below_password_field' ));
+
 		// Available from WordPress 4.3+
 		if ( function_exists( 'wp_site_icon' ) ) {
 			add_action( 'password_protected_login_head', 'wp_site_icon' );
@@ -76,17 +78,21 @@ class Password_Protected {
 
 		add_shortcode( 'password_protected_logout_link', array( $this, 'logout_link_shortcode' ) );
 
-		include_once( dirname( __FILE__ ) . '/admin/admin-bar.php' );
+		include_once dirname( __FILE__ ) . '/admin/admin-bar.php';
 
 		if ( is_admin() ) {
 
-			include_once( dirname( __FILE__ ) . '/admin/admin-caching.php' );
-			include_once( dirname( __FILE__ ) . '/admin/admin.php' );
+			
+			include_once dirname( __FILE__ ) . '/admin/admin-caching.php';
+			include_once dirname( __FILE__ ) . '/admin/admin.php';
 
 			$this->admin_caching = new Password_Protected_Admin_Caching( $this );
-			$this->admin = new Password_Protected_Admin();
+			$this->admin         = new Password_Protected_Admin();
+			
 
 		}
+		include_once dirname( __FILE__ ) . '/admin/class-recaptcha.php';
+		new Password_Protected_reCAPTCHA();
 
 	}
 
@@ -171,7 +177,7 @@ class Password_Protected {
 	/**
 	 * Allow Feeds
 	 *
-	 * @param   boolean  $bool  Allow feeds.
+	 * @param   boolean $bool  Allow feeds.
 	 * @return  boolean         True/false.
 	 */
 	public function allow_feeds( $bool ) {
@@ -187,7 +193,7 @@ class Password_Protected {
 	/**
 	 * Allow Administrators
 	 *
-	 * @param   boolean  $bool  Allow administrators.
+	 * @param   boolean $bool  Allow administrators.
 	 * @return  boolean         True/false.
 	 */
 	public function allow_administrators( $bool ) {
@@ -203,7 +209,7 @@ class Password_Protected {
 	/**
 	 * Allow Users
 	 *
-	 * @param   boolean  $bool  Allow administrators.
+	 * @param   boolean $bool  Allow administrators.
 	 * @return  boolean         True/false.
 	 */
 	public function allow_users( $bool ) {
@@ -221,7 +227,7 @@ class Password_Protected {
 	 *
 	 * If user has a valid email address, return false to disable password protection.
 	 *
-	 * @param   boolean  $bool  Allow IP addresses.
+	 * @param   boolean $bool  Allow IP addresses.
 	 * @return  boolean         True/false.
 	 */
 	public function allow_ip_addresses( $bool ) {
@@ -261,7 +267,7 @@ class Password_Protected {
 	/**
 	 * Encrypt Password
 	 *
-	 * @param  string  $password  Password.
+	 * @param  string $password  Password.
 	 * @return string             Encrypted password.
 	 */
 	public function encrypt_password( $password ) {
@@ -274,8 +280,8 @@ class Password_Protected {
 	 * Maybe Process Logout
 	 */
 	public function maybe_process_logout() {
-
-		if ( isset( $_REQUEST['password-protected'] ) && $_REQUEST['password-protected'] == 'logout' ) {
+		
+		if ( isset( $_REQUEST['password-protected'] ) && sanitize_text_field( $_REQUEST['password-protected'] ) == 'logout' ) {
 
 			$this->logout();
 
@@ -298,41 +304,75 @@ class Password_Protected {
 	public function maybe_process_login() {
 
 		if ( $this->is_active() && isset( $_REQUEST['password_protected_pwd'] ) ) {
-			$password_protected_pwd = $_REQUEST['password_protected_pwd'];
-			$pwd = get_option( 'password_protected_password' );
+			
+			$password_protected_pwd 	= $_REQUEST['password_protected_pwd'];
+			$default_password         	= get_option( 'password_protected_password' );
+			$pwd[]						= $default_password;
 
-			// If correct password...
-			if ( ( hash_equals( $pwd, $this->encrypt_password( $password_protected_pwd ) ) && $pwd != '' ) || apply_filters( 'password_protected_process_login', false, $password_protected_pwd ) ) {
+			$pwd = array_filter( $pwd );
 
-				$remember = isset( $_REQUEST['password_protected_rememberme'] ) ? boolval( $_REQUEST['password_protected_rememberme'] ) : false;
+			$this->errors 				= apply_filters( 'password_protected_verify_recaptcha', $this->errors );
+			if( count( @$this->errors->errors ) > 0 ) return;
+			
+			$pwd = apply_filters( 'password_protected_passwords', $pwd );
+			$auth = false;
 
-				if ( ! $this->allow_remember_me() ) {
-					$remember = false;
+			foreach( $pwd as $i => $p ) {
+				// If correct password...
+				if ( ( hash_equals( $p, $this->encrypt_password( $password_protected_pwd ) ) && $pwd != '' ) || apply_filters( 'password_protected_process_login', false, $password_protected_pwd ) ) {
+					// $auth = true;
+					$auth = apply_filters( 'password_protected_login_password_matched', $p, $this->errors );
+					break;
 				}
-
-				$this->set_auth_cookie( $remember );
-				$redirect_to = isset( $_REQUEST['redirect_to'] ) ? $_REQUEST['redirect_to'] : '';
-				$redirect_to = apply_filters( 'password_protected_login_redirect', $redirect_to );
-
-				if ( ! empty( $redirect_to ) ) {
-					$this->safe_redirect( remove_query_arg( 'password-protected', $redirect_to ) );
-					exit;
-				} elseif ( isset( $_GET['password_protected_pwd'] ) ) {
-					$this->safe_redirect( remove_query_arg( 'password-protected' ) );
-					exit;
-				}
-
-			} else {
-
-				// ... otherwise incorrect password
-				$this->clear_auth_cookie();
-				$this->errors->add( 'incorrect_password', __( 'Incorrect Password', 'password-protected' ) );
-
 			}
-
+			
+			$this->password_protected_process_login( $auth );
+			
 		}
 
 	}
+
+	private function password_protected_process_login( bool $auth ) {
+	
+		if( $auth )
+			$throttle = apply_filters( 'password_protected_check_for_throttling', true );
+
+
+		if( $auth && $throttle ) {
+			
+			do_action( 'password_protected_success_login_attempt' );
+			$remember = isset( $_REQUEST['password_protected_rememberme'] ) ? boolval( $_REQUEST['password_protected_rememberme'] ) : false;
+
+			if ( ! $this->allow_remember_me() ) {
+				$remember = false;
+			}
+			$this->set_auth_cookie( $remember );
+			
+			$redirect_to = isset( $_REQUEST['redirect_to'] ) ? esc_url($_REQUEST['redirect_to']) : '';
+			$redirect_to = apply_filters( 'password_protected_login_redirect', $redirect_to );
+
+			if ( ! empty( $redirect_to ) ) {
+				$this->safe_redirect( remove_query_arg( 'password-protected', $redirect_to ) );
+				exit;
+			} elseif ( isset( $_GET['password_protected_pwd'] ) ) {
+				$this->safe_redirect( remove_query_arg( 'password-protected' ) );
+				exit;
+			} else {
+				$this->safe_redirect( site_url() );
+				exit;
+			}
+		} else {
+			do_action( 'password_protected_failure_login_attempt' );
+
+			// ... otherwise incorrect password
+			$this->clear_auth_cookie();
+			
+			$show_default_error = apply_filters( 'password_protected_throttling_error_messages', true );
+			
+			if( $show_default_error )
+				$this->errors->add( 'incorrect_password', __( 'Incorrect Password', 'password-protected' ) );
+		}
+	} 
 
 	/**
 	 * Is User Logged In?
@@ -350,20 +390,26 @@ class Password_Protected {
 	 */
 	public function maybe_show_login() {
 
-		// Filter for adding exceptions.
-		$show_login = apply_filters( 'password_protected_show_login', $this->is_active() );
-
-		// Logged in
-		if ( $this->is_user_logged_in() )  {
-			$show_login = false;
+		if ( class_exists( 'Login_designer' ) ) {
+			if ( is_customize_preview() ) {
+				return 1;
+			}
 		}
 
+		// Filter for adding exceptions.
+		$show_login = apply_filters( 'password_protected_show_login', $this->is_active() );
+		
+		// Logged in
+		if ( $this->is_user_logged_in() ) {
+			$show_login = false;
+		}
+		
 		if ( ! $show_login ) {
-			return;
+			return 1;
 		}
 
 		// Show login form
-		if ( isset( $_REQUEST['password-protected'] ) && 'login' == $_REQUEST['password-protected'] ) {
+		if ( isset( $_REQUEST['password-protected'] ) && 'login' == sanitize_text_field( $_REQUEST['password-protected'] ) ) {
 
 			$default_theme_file = locate_template( array( 'password-protected-login.php' ) );
 
@@ -432,14 +478,14 @@ class Password_Protected {
 	/**
 	 * Logout URL
 	 *
-	 * @param   string  $redirect_to  Optional. Redirect URL.
+	 * @param   string $redirect_to  Optional. Redirect URL.
 	 * @return  string                Logout URL.
 	 */
 	public function logout_url( $redirect_to = '' ) {
 
 		$query = array(
 			'password-protected' => 'logout',
-			'redirect_to'        => esc_url_raw( $redirect_to )
+			'redirect_to'        => esc_url_raw( $redirect_to ),
 		);
 
 		if ( empty( $query['redirect_to'] ) ) {
@@ -453,7 +499,7 @@ class Password_Protected {
 	/**
 	 * Logout Link
 	 *
-	 * @param   array   $args  Link args.
+	 * @param   array $args  Link args.
 	 * @return  string         HTML link tag.
 	 */
 	public function logout_link( $args = null ) {
@@ -463,10 +509,13 @@ class Password_Protected {
 			return '';
 		}
 
-		$args = wp_parse_args( $args, array(
-			'redirect_to' => '',
-			'text'        => __( 'Logout', 'password-protected' )
-		) );
+		$args = wp_parse_args(
+			$args,
+			array(
+				'redirect_to' => '',
+				'text'        => __( 'Logout', 'password-protected' ),
+			)
+		);
 
 		if ( empty( $args['text'] ) ) {
 			$args['text'] = __( 'Logout', 'password-protected' );
@@ -479,15 +528,19 @@ class Password_Protected {
 	/**
 	 * Logout Link Shortcode
 	 *
-	 * @param   array   $args  Link args.
+	 * @param   array $args  Link args.
 	 * @return  string         HTML link tag.
 	 */
 	public function logout_link_shortcode( $atts, $content = null ) {
 
-		$atts = shortcode_atts( array(
-			'redirect_to' => '',
-			'text'        => $content
-		), $atts, 'logout_link_shortcode' );
+		$atts = shortcode_atts(
+			array(
+				'redirect_to' => '',
+				'text'        => $content,
+			),
+			$atts,
+			'logout_link_shortcode'
+		);
 
 		return $this->logout_link( $atts );
 
@@ -507,8 +560,8 @@ class Password_Protected {
 	/**
 	 * Validate Auth Cookie
 	 *
-	 * @param   string   $cookie  Cookie string.
-	 * @param   string   $scheme  Cookie scheme.
+	 * @param   string $cookie  Cookie string.
+	 * @param   string $scheme  Cookie scheme.
 	 * @return  boolean           Validation successful?
 	 */
 	public function validate_auth_cookie( $cookie = '', $scheme = '' ) {
@@ -529,12 +582,12 @@ class Password_Protected {
 
 		// Quick check to see if an honest cookie has expired
 		if ( $expired < current_time( 'timestamp' ) ) {
-			do_action('password_protected_auth_cookie_expired', $cookie_elements);
+			do_action( 'password_protected_auth_cookie_expired', $cookie_elements );
 			return false;
 		}
 
-		$key = md5( $this->get_site_id() . $this->get_hashed_password() . '|' . $expiration );
-		$hash = hash_hmac( 'md5', $this->get_site_id() . '|' . $expiration, $key);
+		$key  = md5( $this->get_site_id() . $this->get_hashed_password() . '|' . $expiration );
+		$hash = hash_hmac( 'md5', $this->get_site_id() . '|' . $expiration, $key );
 
 		if ( $hmac != $hash ) {
 			do_action( 'password_protected_auth_cookie_bad_hash', $cookie_elements );
@@ -552,14 +605,14 @@ class Password_Protected {
 	/**
 	 * Generate Auth Cookie
 	 *
-	 * @param   int     $expiration  Expiration time in seconds.
-	 * @param   string  $scheme      Cookie scheme.
+	 * @param   int    $expiration  Expiration time in seconds.
+	 * @param   string $scheme      Cookie scheme.
 	 * @return  string               Cookie.
 	 */
 	public function generate_auth_cookie( $expiration, $scheme = 'auth' ) {
 
-		$key = md5( $this->get_site_id() . $this->get_hashed_password() . '|' . $expiration );
-		$hash = hash_hmac( 'md5', $this->get_site_id() . '|' . $expiration, $key );
+		$key    = md5( $this->get_site_id() . $this->get_hashed_password() . '|' . $expiration );
+		$hash   = hash_hmac( 'md5', $this->get_site_id() . '|' . $expiration, $key );
 		$cookie = $this->get_site_id() . '|' . $expiration . '|' . $hash;
 
 		return $cookie;
@@ -569,8 +622,8 @@ class Password_Protected {
 	/**
 	 * Parse Auth Cookie
 	 *
-	 * @param   string  $cookie  Cookie string.
-	 * @param   string  $scheme  Cookie scheme.
+	 * @param   string $cookie  Cookie string.
+	 * @param   string $scheme  Cookie scheme.
 	 * @return  string           Cookie string.
 	 */
 	public function parse_auth_cookie( $cookie = '', $scheme = '' ) {
@@ -604,18 +657,18 @@ class Password_Protected {
 	 *
 	 * @todo
 	 *
-	 * @param  boolean  $remember  Remember logged in.
-	 * @param  string   $secure    Secure cookie.
+	 * @param  boolean $remember  Remember logged in.
+	 * @param  string  $secure    Secure cookie.
 	 */
-	public function set_auth_cookie( $remember = false, $secure = '') {
+	public function set_auth_cookie( $remember = false, $secure = '' ) {
 
 		if ( $remember ) {
 			$expiration_time = apply_filters( 'password_protected_auth_cookie_expiration', get_option( 'password_protected_remember_me_lifetime', 14 ) * DAY_IN_SECONDS, $remember );
-			$expiration = $expire = current_time( 'timestamp' ) + $expiration_time;
+			$expiration      = $expire = current_time( 'timestamp' ) + $expiration_time;
 		} else {
 			$expiration_time = apply_filters( 'password_protected_auth_cookie_expiration', DAY_IN_SECONDS * 20, $remember );
-			$expiration = current_time( 'timestamp' ) + $expiration_time;
-			$expire = 0;
+			$expiration      = current_time( 'timestamp' ) + $expiration_time;
+			$expire          = 0;
 		}
 
 		if ( '' === $secure ) {
@@ -623,7 +676,7 @@ class Password_Protected {
 		}
 
 		$secure_password_protected_cookie = apply_filters( 'password_protected_secure_password_protected_cookie', false, $secure );
-		$password_protected_cookie = $this->generate_auth_cookie( $expiration, 'password_protected' );
+		$password_protected_cookie        = $this->generate_auth_cookie( $expiration, 'password_protected' );
 
 		setcookie( $this->cookie_name(), $password_protected_cookie, $expire, COOKIEPATH, COOKIE_DOMAIN, $secure_password_protected_cookie, true );
 		if ( COOKIEPATH != SITECOOKIEPATH ) {
@@ -664,7 +717,7 @@ class Password_Protected {
 		$old_version = get_option( 'password_protected_version' );
 
 		// 1.1 - Upgrade to MD5
-		if ( empty( $old_version ) || version_compare( '1.1', $old_version ) ) {
+		if ( empty( $old_version ) || $old_version == '1.1' ) {
 			$pwd = get_option( 'password_protected_password' );
 			if ( ! empty( $pwd ) ) {
 				$new_pwd = $this->encrypt_password( $pwd );
@@ -689,7 +742,7 @@ class Password_Protected {
 		if ( class_exists( 'CWS_Login_Logo_Plugin' ) ) {
 
 			// Add support for Mark Jaquith's Login Logo plugin
-			add_action( 'password_protected_login_head', array( new CWS_Login_Logo_Plugin, 'login_head' ) );
+			add_action( 'password_protected_login_head', array( new CWS_Login_Logo_Plugin(), 'login_head' ) );
 
 		} elseif ( class_exists( 'UberLoginLogo' ) ) {
 
@@ -714,7 +767,7 @@ class Password_Protected {
 
 		if ( $this->errors->get_error_code() ) {
 
-			$errors = '';
+			$errors   = '';
 			$messages = '';
 
 			foreach ( $this->errors->get_error_codes() as $code ) {
@@ -734,7 +787,6 @@ class Password_Protected {
 			if ( ! empty( $messages ) ) {
 				echo '<p class="message">' . apply_filters( 'password_protected_login_messages', $messages ) . "</p>\n";
 			}
-
 		}
 
 	}
@@ -759,14 +811,13 @@ class Password_Protected {
 		if ( ! empty( $located ) ) {
 
 			$stylesheet_directory = trailingslashit( get_stylesheet_directory() );
-			$template_directory = trailingslashit( get_template_directory() );
+			$template_directory   = trailingslashit( get_template_directory() );
 
 			if ( $stylesheet_directory == substr( $located, 0, strlen( $stylesheet_directory ) ) ) {
 				wp_enqueue_style( 'password-protected-login', get_stylesheet_directory_uri() . '/' . $filename );
-			} else if ( $template_directory == substr( $located, 0, strlen( $template_directory ) ) ) {
+			} elseif ( $template_directory == substr( $located, 0, strlen( $template_directory ) ) ) {
 				wp_enqueue_style( 'password-protected-login', get_template_directory_uri() . '/' . $filename );
 			}
-
 		}
 
 	}
@@ -807,7 +858,7 @@ class Password_Protected {
 	 * Always allow logged in users who require REST API for Gutenberg
 	 * and other admin/plugin compatibility.
 	 *
-	 * @param   WP_REST_Request   $access  Full details about the request.
+	 * @param   WP_REST_Request $access  Full details about the request.
 	 * @return  WP_Error|boolean
 	 */
 	public function only_allow_logged_in_rest_access( $access ) {
@@ -819,6 +870,26 @@ class Password_Protected {
 
 		return $access;
 
+	}
+
+	/**
+	 * Print text above password field
+	 * @return  void.
+	 */
+	public function password_protected_above_password_field() {
+		$text = get_option('password_protected_text_above_password');
+		if( !empty($text) )
+			echo '<div class="password-protected-text-above">' . esc_attr( $text ) . '</div>'; 
+	}
+
+	/**
+	 * Print text below password field
+	 * @return  void.
+	 */
+	public function password_protected_below_password_field() {
+		$text = get_option('password_protected_text_below_password');
+		if( !empty($text) )
+			echo '<div class="password-protected-text-below">' . esc_attr( $text ) . '</div>'; 
 	}
 
 }