+++ theme/password-protected-login.php
@@ -90,6 +90,8 @@ if ( version_compare( $wp_version, '3.9-dev', '>=' ) ) {
 
 <style media="screen">
 #login_error, .login .message, #loginform { margin-bottom: 20px; }
+.password-protected-text-below { display: inline-block; text-align: center; margin-top: 30px;}
+.password-protected-text-above { text-align: center; margin-bottom: 10px;}
 </style>
 
 <?php
@@ -107,6 +109,9 @@ if ( $is_iphone ) {
 }
 
 do_action( 'login_enqueue_scripts' );
+if ( class_exists( 'Login_Designer' ) ) {
+	do_action( 'password_protected_enqueue_scripts' );
+}
 do_action( 'password_protected_login_head' );
 
 ?>
@@ -115,29 +120,31 @@ do_action( 'password_protected_login_head' );
 <body class="login login-password-protected login-action-password-protected-login wp-core-ui">
 
 <div id="login">
-	<h1><a href="<?php echo esc_url( apply_filters( 'password_protected_login_headerurl', home_url( '/' ) ) ); ?>" title="<?php echo esc_attr( apply_filters( 'password_protected_login_headertitle', get_bloginfo( 'name' ) ) ); ?>"><?php bloginfo( 'name' ); ?></a></h1>
-
+	<h1 id="password-protected-logo"><a href="<?php echo esc_url( apply_filters( 'password_protected_login_headerurl', home_url( '/' ) ) ); ?>" title="<?php echo esc_attr( apply_filters( 'password_protected_login_headertitle', get_bloginfo( 'name' ) ) ); ?>"><?php bloginfo( 'name' ); ?></a></h1>
 	<?php do_action( 'password_protected_login_messages' ); ?>
 	<?php do_action( 'password_protected_before_login_form' ); ?>
 
 	<form name="loginform" id="loginform" action="<?php echo esc_url( $Password_Protected->login_url() ); ?>" method="post">
+		<?php do_action( 'password_protected_above_password_field' ); ?>
 		<p>
-			<label for="password_protected_pass"><?php echo apply_filters( 'password_protected_login_password_title', __( 'Password', 'password-protected' ) ); ?><br />
-			<input type="password" name="password_protected_pwd" id="password_protected_pass" class="input" value="" size="20" tabindex="20" /></label>
+			<label for="password_protected_pass"><?php echo esc_attr( apply_filters( 'password_protected_login_password_title', __( 'Password', 'password-protected' ) ) ); ?></label>
+			<input type="password" name="password_protected_pwd" id="password_protected_pass" class="input" value="" size="20" tabindex="20" />
 		</p>
-
+		<?php do_action('password_protected_after_password_field'); ?>
 		<?php if ( $Password_Protected->allow_remember_me() ) : ?>
 			<p class="forgetmenot">
 				<label for="password_protected_rememberme"><input name="password_protected_rememberme" type="checkbox" id="password_protected_rememberme" value="1" tabindex="90" /> <?php esc_attr_e( 'Remember Me' ); ?></label>
 			</p>
 		<?php endif; ?>
-
+		
 		<p class="submit">
 			<input type="submit" name="wp-submit" id="wp-submit" class="button button-primary button-large" value="<?php esc_attr_e( 'Log In' ); ?>" tabindex="100" />
 			<input type="hidden" name="password_protected_cookie_test" value="1" />
 			<input type="hidden" name="password-protected" value="login" />
-			<input type="hidden" name="redirect_to" value="<?php echo esc_attr( ! empty( $_REQUEST['redirect_to'] ) ? $_REQUEST['redirect_to'] : '' ); ?>" />
+			<input type="hidden" name="redirect_to" value="<?php echo esc_attr( ! empty( $_REQUEST['redirect_to'] ) ? esc_url( $_REQUEST['redirect_to'] ) : '' ); ?>" />
 		</p>
+		<?php do_action( 'password_protected_below_password_field' ); ?>
+		
 	</form>
 
 	<?php do_action( 'password_protected_after_login_form' ); ?>
@@ -152,6 +159,10 @@ if(typeof wpOnload=='function')wpOnload();
 <?php do_action( 'login_footer' ); ?>
 
 <div class="clear"></div>
+
+<?php if ( class_exists( 'Login_Designer' ) ) : ?>
+	<div id="password-protected-background" style="position:absolute; inset: 0;width: 100%;height: 100%;z-index: -1;transition: opacity 300ms cubic-bezier(0.694, 0, 0.335, 1) 0s"></div>
+<?php endif; ?>
 
 </body>
 </html>