+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.5.3 - 2022-06-17
+- Improved Settings HTML structure
+- Added Note regarding compatibility with login designer within dashboard
+
+## 2.5.2 - 2022-05-26
+Made compatibility with [login designer](https://wp.org/plugins/login-designer); Now you can customize the password-protected screen with the customizer using login designer plugin.
+
 ## [2.5] - 2021-09-01
 
 ### Fixed